Added checkpoint checks for execution

+++ src/Modules/Tours/Explorer.Tours.API/Dtos/LocationCreateDto.cs
@@ -9,7 +9,7 @@ namespace Explorer.Tours.API.Dtos
     public class LocationCreateDto
     {
         public long? Id { get; set; }
-        public long Latitude { get; set; }
-        public long Longitude { get; set; }
+        public double  Latitude { get; set; }
+        public double Longitude { get; set; }
     }
 }
